+++ src/pages/top/components/FAQContainer/index.module.scss
