+++ src/pages/top/components/FAQContainer/index.tsx
@@ -1,8 +1,8 @@
 import React, { FC } from 'react';
 import styles from './index.module.scss';
-import { FAQList } from '@/pages/topa/components/FAQContainer/const/FAQList';
-import { useFAQ } from '@/pages/topa/hooks/useFAQ';
-import { FAQItem } from '@/pages/topa/components/FAQItem';
+import { FAQList } from '@/pages/top/components/FAQContainer/const/FAQList';
+import { useFAQ } from '@/pages/top/hooks/useFAQ';
+import { FAQItem } from '@/pages/top/components/FAQItem';
 
 export const FAQContainer: FC = () => {
   const { isAnswerVisible, getActiveIndex } = useFAQ();
