+++ src/pages/top/components/FAQItem/index.module.scss
