+++ src/pages/top/components/PriceSection/index.module.scss
