+++ src/pages/top/components/TopViewAutoImageSlide/index.stories.ts
