+++ src/pages/top/components/VerticalText/index.module.scss
