+++ src/pages/top/components/VerticalText/index.stories.tsx
