+++ src/pages/top/components/VerticalText/index.tsx
