+++ src/pages/top/components/WorksSlider/const/distanceToHero.ts
