+++ src/pages/top/components/WorksSlider/const/worksSlideList.ts
