+++ src/pages/top/components/WorksSlider/index.tsx
@@ -1,15 +1,15 @@
 import React, { FC } from 'react';
 import { useActiveIndex } from '@/hooks/useActiveIndex';
-import { useWorksSliderBreakPointsStyle } from '@/pages/topa/hooks/useWorksSliderBreakPointsStyle';
+import { useWorksSliderBreakPointsStyle } from '@/pages/top/hooks/useWorksSliderBreakPointsStyle';
 import { useAnimationFrameInterval } from '@/hooks/useAnimationFrameInterval';
-import { worksSlideList } from '@/pages/topa/components/WorksSlider/const/worksSlideList';
-import styles from '@/pages/topa/components/WorksSlider/index.module.scss';
-import { generateSlideStyle } from '@/pages/topa/logics/getDynamicSliderStyle';
+import { worksSlideList } from '@/pages/top/components/WorksSlider/const/worksSlideList';
+import styles from '@/pages/top/components/WorksSlider/index.module.scss';
+import { generateSlideStyle } from '@/pages/top/logics/getDynamicSliderStyle';
 import Image from 'next/image';
 import { WorksHeroImageDescription } from '../WorksHeroImageDescription';
-import { isHeroImage } from '@/pages/topa/logics/isHeroImage';
-import { getLeftIndex } from '@/pages/topa/logics/getLeftIndex';
-import { DISTANCE_TO_HERO } from '@/pages/topa/components/WorksSlider/const/distanceToHero';
+import { isHeroImage } from '@/pages/top/logics/isHeroImage';
+import { getLeftIndex } from '@/pages/top/logics/getLeftIndex';
+import { DISTANCE_TO_HERO } from '@/pages/top/components/WorksSlider/const/distanceToHero';
 
 export const WorksSlider: FC = () => {
   const [activeIndex, setActiveIndex] = useActiveIndex();
