+++ src/pages/top/logics/getDynamicSliderStyle/fixtures.ts
