+++ src/pages/top/logics/getDynamicSliderStyle/index.ts
@@ -1,6 +1,6 @@
 import React from 'react';
 import { StaticImageData } from 'next/image';
-import { UseWorksSliderBreakPoints } from '@/pages/topa/hooks/useWorksSliderBreakPointsStyle';
+import { UseWorksSliderBreakPoints } from '@/pages/top/hooks/useWorksSliderBreakPointsStyle';
 
 export type BaseSlideList = {
   id: string;
