+++ src/pages/top/logics/getLeftIndex/fixtures.ts
