+++ src/pages/top/logics/isHeroImage/fixtures.ts
